requirements.txt
aiohttp==3.9.4
fastapi==0.112.0
httpx==0.27.0
motor==3.4.0
pymongo==4.6.3
python-dotenv==1.0.1
SimplerLLM==0.3.0.3
SQLAlchemy==2.0.29
aiomysql==0.2.0